- blake commit

+++ 1503-StructByLightning-LostSamurai-StrawManProject/Lost Samurai/Last Samurai/Arrow.cpp
@@ -40,7 +40,7 @@ void Arrow::Render()
 	//draw rect
 	//SGD::GraphicsManager::GetInstance()->DrawRectangle(draw, SGD::Color{ 255, 0, 0, 0 }, {}, 3);
 	//SGD::GraphicsManager::GetInstance()->DrawTexture(GetImage(), pt, {}, {}, {}, SGD::Size{ 0.7f, 0.7f });
-	AnimationSystem::GetInstance()->Render(m_Timestamp, (int)pt.x, (int)pt.y, SGD::Size{ -.3f, -.3f });
+	AnimationSystem::GetInstance()->Render(m_Timestamp, pt.x, pt.y, SGD::Size{ -.3f, -.3f });
 	/*if (Debug)
 	{*/
 	SGD::Rectangle re = GetRect();
